Remove identity screen
