+++ fe1-web/src/core/navigation/typing/LaoParamList.ts
@@ -8,9 +8,8 @@ import { SocialParamList } from './SocialParamList';
 import { WalletParamList } from './WalletParamList';
 
 export type LaoParamList = {
-  [STRINGS.navigation_lao_home]: undefined;
-  [STRINGS.navigation_lao_notifications]: NavigatorScreenParams<NotificationParamList>;
   [STRINGS.navigation_lao_events]: NavigatorScreenParams<LaoEventsParamList>;
+  [STRINGS.navigation_lao_notifications]: NavigatorScreenParams<NotificationParamList>;
   [STRINGS.navigation_social_media]: NavigatorScreenParams<SocialParamList>;
   [STRINGS.navigation_lao_wallet]: NavigatorScreenParams<WalletParamList>;
 };
