+++ fe1-web/src/features/home/screens/ConnectConfirm.tsx
@@ -64,7 +64,8 @@ const ConnectConfirm = () => {
       }
 
       navigation.navigate(STRINGS.navigation_app_lao, {
-        screen: STRINGS.navigation_lao_home,
+        screen: STRINGS.navigation_lao_events,
+        params: { screen: STRINGS.navigation_lao_events_home },
       });
     } catch (err) {
       console.error(`Failed to establish lao connection`, err);
