+++ fe1-web/src/features/home/screens/ConnectScan.tsx
@@ -185,7 +185,8 @@ const ConnectScan = () => {
       setIsConnecting(false);
 
       navigation.navigate(STRINGS.navigation_app_lao, {
-        screen: STRINGS.navigation_lao_home,
+        screen: STRINGS.navigation_lao_events,
+        params: { screen: STRINGS.navigation_lao_events_home },
       });
     } catch (error) {
       // close already established connections
