+++ fe1-web/src/features/home/screens/Launch.tsx
@@ -63,7 +63,8 @@ const Launch = () => {
 
       // navigate to the newly created LAO
       navigation.navigate(STRINGS.navigation_app_lao, {
-        screen: STRINGS.navigation_lao_home,
+        screen: STRINGS.navigation_lao_events,
+        params: { screen: STRINGS.navigation_lao_events_home },
       });
     } catch (e) {
       console.error(`Failed to establish lao connection`, e);
