+++ fe1-web/src/features/lao/screens/IdentityScreen.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import ScreenWrapper from 'core/components/ScreenWrapper';
+
+import Identity from '../components/Identity';
+
+const IdentityScreen = () => {
+  return (
+    <ScreenWrapper>
+      <Identity />
+    </ScreenWrapper>
+  );
+};
+
+export default IdentityScreen;
