+++ fe1-web/src/features/lao/screens/__tests__/EventsScreen.test.tsx
@@ -12,7 +12,11 @@ import { encodeLaoConnectionForQRCode } from 'features/home/functions';
 import { LAO_FEATURE_IDENTIFIER, LaoReactContext } from 'features/lao/interface';
 import { laoReducer, setCurrentLao } from 'features/lao/reducer';
 
-import EventsScreen from '../EventsScreen';
+import EventsScreen, {
+  EventsScreenHeader,
+  EventsScreenHeaderLeft,
+  EventsScreenHeaderRight,
+} from '../EventsScreen';
 
 const contextValue = {
   [LAO_FEATURE_IDENTIFIER]: {
@@ -34,12 +38,51 @@ const mockStore = configureStore({
 mockStore.dispatch(setCurrentLao(mockLao.toState()));
 mockStore.dispatch(setKeyPair(mockPopToken.toState()));
 
-describe('AttendeeScreen', () => {
+describe('EventsScreen', () => {
   it('renders correctly', () => {
     const component = render(
       <Provider store={mockStore}>
         <FeatureContext.Provider value={contextValue}>
           <MockNavigator component={EventsScreen} />
+        </FeatureContext.Provider>
+      </Provider>,
+    ).toJSON();
+    expect(component).toMatchSnapshot();
+  });
+});
+
+describe('EventsScreenHeader', () => {
+  it('renders correctly', () => {
+    const component = render(
+      <Provider store={mockStore}>
+        <FeatureContext.Provider value={contextValue}>
+          <MockNavigator component={EventsScreenHeader} />
+        </FeatureContext.Provider>
+      </Provider>,
+    ).toJSON();
+    expect(component).toMatchSnapshot();
+  });
+});
+
+describe('EventsScreenHeaderLeft', () => {
+  it('renders correctly', () => {
+    const component = render(
+      <Provider store={mockStore}>
+        <FeatureContext.Provider value={contextValue}>
+          <MockNavigator component={EventsScreenHeaderLeft} />
+        </FeatureContext.Provider>
+      </Provider>,
+    ).toJSON();
+    expect(component).toMatchSnapshot();
+  });
+});
+
+describe('EventsScreenHeaderRight', () => {
+  it('renders correctly', () => {
+    const component = render(
+      <Provider store={mockStore}>
+        <FeatureContext.Provider value={contextValue}>
+          <MockNavigator component={EventsScreenHeaderRight} />
         </FeatureContext.Provider>
       </Provider>,
     ).toJSON();
