--- fe1-web/src/features/lao/screens/index.ts
@@ -1,1 +1,0 @@
-export { default as EventsScreen } from 'features/lao/screens/EventsScreen';
