+++ fe1-web/src/resources/strings.ts
@@ -107,8 +107,6 @@ namespace STRINGS {
   export const lao_properties_id = 'Identifier';
   export const lao_properties_your_role = 'Your role';
   export const lao_properties_current_connections = 'You are currently connected to';
-  export const lao_properties_add_additional_connection = 'Add connection';
-  export const lao_properties_disconnect = 'Disconnect';
   export const lao_no_current = 'You are currently not connected to any LAO';
   export const lao_error_disconnect = 'The connection to the LAO broke unexpectedly';
